lung-segmentation-app/requirements.txt
# Lung Segmentation Application - Requirements

# Core ML/DL Libraries
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Web Application
streamlit>=1.28.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
PyYAML>=6.0
pathlib>=1.0.1
tqdm>=4.65.0

# Image Processing
albumentations>=1.3.0
imgaug>=0.4.0

# Metrics & Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Annotation Tools
pycocotools>=2.0.6

# Optional but recommended
# GPU acceleration (if CUDA available)
# nvidia-cuda-runtime-cu11>=11.8
# nvidia-cudnn-cu11>=8.6
